+++ app/Http/Controllers/Backend/Product/UnitController.php
@@ -14,6 +14,7 @@ class UnitController extends Controller
      */
     public function index(Request $request)
     {
+        abort_if(!auth()->user()->can('unit_view'), 403);
         if ($request->ajax()) {
             $units = Unit::latest()->get();
             return DataTables::of($units)
@@ -47,7 +48,8 @@ class UnitController extends Controller
      * Show the form for creating a new resource.
      */
     public function create()
-    {;
+    {
+        abort_if(!auth()->user()->can('unit_create'), 403);
         return view('backend.units.create');
     }
 
@@ -56,6 +58,7 @@ class UnitController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(!auth()->user()->can('unit_create'), 403);
         $unit = Unit::create($request->only(['title','short_name']));
 
         return redirect()->route('backend.admin.units.index')->with('success', 'Unit created successfully!');
@@ -74,6 +77,7 @@ class UnitController extends Controller
      */
     public function edit($id)
     {
+        abort_if(!auth()->user()->can('unit_update'), 403);
 
         $unit = Unit::findOrFail($id);
         return view('backend.units.edit', compact('unit'));
@@ -84,6 +88,7 @@ class UnitController extends Controller
      */
     public function update(Request $request, $id)
     {
+        abort_if(!auth()->user()->can('unit_update'), 403);
         $unitToUpdate = Unit::findOrFail($id);
         $unitToUpdate->update($request->only(['title', 'short_name']));
         return redirect()->route('backend.admin.units.index')->with('success', 'Unit updated successfully!');
@@ -95,6 +100,7 @@ class UnitController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(!auth()->user()->can('unit_delete'), 403);
         $unit = Unit::findOrFail($id);
         $unit->delete();
         return redirect()->back()->with('success', 'Unit Deleted Successfully');
