+++ app/Http/Controllers/Backend/SupplierController.php
@@ -14,6 +14,7 @@ class SupplierController extends Controller
      */
     public function index(Request $request)
     {
+    abort_if(!auth()->user()->can('supplier_view'), 403);
         if ($request->ajax()) {
             $suppliers = Supplier::latest()->get();
             return DataTables::of($suppliers)
@@ -56,6 +57,7 @@ class SupplierController extends Controller
      */
     public function create()
     {
+    abort_if(!auth()->user()->can('supplier_create'), 403);
         return view('backend.suppliers.create');
     }
 
@@ -65,6 +67,7 @@ class SupplierController extends Controller
     public function store(Request $request)
     {
 
+    abort_if(!auth()->user()->can('supplier_create'), 403);
         if ($request->wantsJson()) {
             $request->validate([
                 'name' => 'required|string',
@@ -101,6 +104,7 @@ class SupplierController extends Controller
      */
     public function edit($id)
     {
+        abort_if(!auth()->user()->can('supplier_update'), 403);
         $supplier = Supplier::findOrFail($id);
         return view('backend.suppliers.edit', compact('supplier'));
     }
@@ -110,6 +114,7 @@ class SupplierController extends Controller
      */
     public function update(Request $request, $id)
     {
+        abort_if(!auth()->user()->can('supplier_update'), 403);
         $supplier = Supplier::findOrFail($id);
 
         $request->validate([
@@ -130,6 +135,7 @@ class SupplierController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(!auth()->user()->can('supplier_delete'), 403);
         $supplier = Supplier::findOrFail($id);
         $supplier->delete();
         session()->flash('success', 'Supplier deleted successfully.');
